==> gazdasag/2019/3_formatter.ipynb <==
import requests, pandas as pd, numpy as np
from requests import session
from bs4 import BeautifulSoup

manual=pd.read_excel('manual_manual.xlsx').set_index('Cégnév')
data=pd.read_excel('data.xlsx').set_index('Cégnév')
sectors=pd.read_excel('valid_manual.xlsx')

len(data), len(manual)

data=pd.concat([data,manual])
data['Sector']=sectors.set_index(0)[1]

Bring in 2014 data

manual4=pd.read_excel('../2018/manual_manual.xlsx').set_index('Cégnév')
data4=pd.read_excel('../2018/data.xlsx').set_index('Cégnév')
sectors4=pd.read_excel('../2018/valid_manual.xlsx')

len(data4), len(manual4)

data4=pd.concat([data4,manual4])
data4['Sector4']=sectors4.set_index(0)[1]

data=data.join(data4[['Alkalmazottak száma 2014','Árbevétel 2014 (RON)']])

repl={'udvarhelyszek':'Udvarhelyszék',
'csikszek':'Csíkszék',
'gyergyoszek':'Gyergyószék',
'marosszek':'Marosszék',
'also-haromszek':'Alsó-háromszék',
'felso-haromszek':'Felső-háromszék'}
for r in repl:
    data=data.replace(r,repl[r])

data.to_excel('export1.xlsx')

d1=data[['Régió', 'Latitude', 'Longitude', 'Sector']]
d1.columns=['Régió', 'Hosszúság', 'Szélesség', 'Iparág']

d1.to_excel('d1.xlsx')

data.columns

ds=[]
ds1=[]
ds2=[]
for i in range(2014,2019):
    g='Alkalmazottak száma'
    c=g+' '+str(i)
    print(c)
    df=data[[c]]
    df.columns=[g]
    df[g]=df[g].fillna('').str.replace(u'\xa0','').str.replace('-','')
    df['Év']=i
    ds1.append(df)
    ds.append(d1.join(df))
    g='Árbevétel'
    c=g+' '+str(i)+' (RON)'
    print(c)
    df=data[[c]]
    df.columns=[g]
    df[g]=df[g].fillna('').str.replace(u'\xa0','').str.replace('-','')
    df['Év']=i
    ds2.append(df)

df=pd.concat(ds,sort=False)

df1=pd.concat(ds1,sort=False)
df2=pd.concat(ds2,sort=False)

df1.to_excel('df1.xlsx')
df2.to_excel('df2.xlsx')

df1=df1.set_index('Év',append=True)
df2=df2.set_index('Év',append=True)

df=df1.join(df2).replace('',np.nan).astype(float)

df['Bevétel']=np.round(df['Árbevétel']/1000000.0,3)

df['Bevétel/Alkalmazott']=np.round(df['Bevétel']/df['Alkalmazottak száma'],3)

data2=df.reset_index().join(d1,on='Cégnév').set_index('Cégnév')

Normalize names

#[' '.join([j.title() if len(j)>3 else j for j in i.split(' ')]) for i in data2.index.str.replace(' SA','').str.replace(' SRL','').str.replace(' ROMANIA ','')]
data2.index=data2.index.str.replace(' SA','').str.replace(' SRL','').str.replace(' ROMANIA','')

data2.to_excel('export2.xlsx')

Calculate percentages

totals=data2.groupby(['Régió','Év']).sum()[['Alkalmazottak száma','Bevétel']]
totals.columns=['TAlkalmazottak száma','TBevétel']
totals=totals.unstack()
totals=(np.round(totals/totals.sum()*100,0)).stack().astype(str)
totals['TAlkalmazottak száma']=totals['TAlkalmazottak száma'].str[:-2]+'%'
totals['TBevétel']=totals['TBevétel'].str[:-2]+'%'

totals2=data2.groupby(['Cégnév','Év']).sum()[['Alkalmazottak száma','Bevétel']]
totals2.columns=['TAlkalmazottak száma','TBevétel']
totals2=totals2.unstack()
totals2=(np.round(totals2/totals2.sum()*100,1)).stack().astype(str)
totals2['TAlkalmazottak száma']=totals2['TAlkalmazottak száma'].str[:]+'%'
totals2['TBevétel']=totals2['TBevétel'].str[:]+'%'
totals2.columns=['T2Alkalmazottak száma','T2Bevétel']

totals2.loc['AZOMURES']

totals3=data2.groupby(['Cégnév','Év','Régió']).sum()[['Alkalmazottak száma','Bevétel']]
totals3.columns=['TAlkalmazottak száma','TBevétel']
totals3=totals3.unstack().unstack()
totals3=(np.round(totals3/totals3.sum()*100,1)).stack().stack().astype(str)
totals3['TAlkalmazottak száma']=totals3['TAlkalmazottak száma'].str[:]+'%'
totals3['TBevétel']=totals3['TBevétel'].str[:]+'%'
totals3.columns=['T3Alkalmazottak száma','T3Bevétel']

totals3.loc['AZOMURES']

data22=data2.reset_index().set_index(['Régió','Év']).join(totals).reset_index()
data22=data22.set_index(['Cégnév','Év']).join(totals2).reset_index()
data22=data22.set_index(['Cégnév','Év','Régió']).join(totals3).reset_index()

data221=data22[['Régió','Év','Bevétel','TBevétel','T2Bevétel','T3Bevétel','Cégnév']]
data221['Régió %']=data221['TBevétel']+' '+data221['Régió']
data221['Cégnév %']=data221['Cégnév']+' | '+data221['T2Bevétel']+' | megyében '+data221['T3Bevétel']
data221=data221.set_index(['Cégnév %','Régió %','Év'])['Bevétel'].unstack()
data221['Type']='Árbevétel (millió RON)'
data221=data221.reset_index()

data222=data22[['Régió','Év','Alkalmazottak száma','TAlkalmazottak száma','T2Alkalmazottak száma','T3Alkalmazottak száma','Cégnév']]
data222['Régió %']=data222['TAlkalmazottak száma']+' '+data222['Régió']
data222['Cégnév %']=data222['Cégnév']+' | ' +data222['T2Alkalmazottak száma']+' | megyében '+data222['T3Alkalmazottak száma']
data222=data222.set_index(['Cégnév %','Régió %','Év'])['Alkalmazottak száma'].unstack()
data222['Type']='Alkalmazottak száma'
data222=data222.reset_index()

data223=pd.concat([data221,data222])

data223.to_excel('export22.xlsx')

colors22={'Alsó-háromszék': '#855c75',
'Csíkszék': '#526a83',
'Felső-háromszék': '#625377',
'Gyergyószék': '#af6458',
'Marosszék': '#736f4c',
'Udvarhelyszék': '#d9af6b'}

for i in data223['Régió %'].unique():
    print(i,':',colors22[i[i.find(' ')+1:]])

Simplify

data3=df.unstack()
data3.columns=data3.columns.get_level_values(1).astype(str)
data3=data3.reset_index().join(d1,on='Cégnév').set_index('Cégnév')
data3.index.name=None
data3.index=data3.index.str.replace(' SA','').str.replace(' SRL','').str.replace(' ROMANIA','')
data3.to_excel('export3.xlsx')

data3b=data3.copy()
data3b.columns=range(len(data3b.columns))
data3b=data3b[[14,20,21,22,23]]
data3b['Size']=data3b[14]**0.5+5
data3b.dropna(subset=[14]).to_excel('export3b.xlsx')

data4=data3.groupby(['Iparág','Régió']).sum()
data4b=data3.groupby(['Régió']).sum()
data4=data4.unstack().T
data4b=data4b.unstack().T
data4['Mind']=data4b
data4=data4.unstack().T
data4.reset_index().to_excel('export4.xlsx')

data4=data3.groupby(['Régió','Iparág']).sum()
data4b=data3.groupby(['Iparág']).sum()
data4=data4.unstack().T
data4b=data4b.unstack().T
data4['Mind']=data4b
data4=data4.unstack().T
data4.reset_index().to_excel('export4b.xlsx')

d1s=[]
for y in ['2014','2015','2016','2017','2018']:
    d1=data4[[y]]
    d1.columns=[0,1,'Bevétel',2]
    d1=d1[['Bevétel']]
    d1['Év']=y
    d1s.append(d1)
d1s=pd.concat(d1s)
d1s.reset_index().to_excel('export5.xlsx')

d2s=d1s.reset_index().set_index(['Iparág','Év','Régió']).unstack()
d3s=d2s.reset_index().groupby(['Év']).sum()
d3s['Iparág']='Mind'
d3s=d3s.reset_index().set_index(['Iparág','Év'])
d6=pd.concat([d2s,d3s])['Bevétel']
d6.reset_index().to_excel('export6.xlsx')

d6T=d6.stack().reset_index().set_index(['Régió','Év','Iparág']).unstack().reset_index()

d6T[d6T['Régió']!='Mind'].to_excel('export6T.xlsx')


regiok=[i for i in d6.columns if i!='Mind']
d6=d6.reset_index()

for i in regiok:
    d6[i]=np.round(d6[i]*100/d6['Mind'],0)

d6=d6.replace(0,np.nan).replace(1,np.nan).replace(2,np.nan)

d6.to_excel('export6b.xlsx')
